重构1
