+++ restdemo/resource/hello.py
@@ -1,0 +1,5 @@
+from flask_restful import Resource
+
+class HelloWorld(Resource):
+    def get(self):
+        return 'hello world'
